// src/features/product/types/product.enums.ts
export enum ProductIsHotEnum {
  HOT = 1,
  NORMAL = 2,
}

export enum ProductStatusEnum {
  DISPLAY = 1,
  HIDDEN = 2,
}

export enum LayoutsEnum {
  LIST = "list",
  GRID = "grid",
}

export enum SortQueryEnum {
  NAME_ASC = "name_asc",
  NAME_DESC = "name_desc",
  PRICE_ASC = "price_asc",
  PRICE_DESC = "price_desc",
  CREATED_AT_ASC = "created_at_asc",
  CREATED_AT_DESC = "created_at_desc",
}